+++ internal/pkg/model/models/dumb/Model.go
@@ -159,3 +159,5 @@ func (m *Model) DecisionVariableChange(decisionVariableName string) float64 {
 }
 
 func (m *Model) ChangeIsValid() (bool, *errors.CompositeError) { return true, nil }
+
+func (m *Model) IsEquivalentTo(model.Model) bool { return false } // TODO: implement